refactor(#3038): move tmux watcher liveness decisions
